+++ src/Network/Wai/SAML2/Validation.hs
@@ -30,7 +30,6 @@ import qualified Data.ByteString as BS
 import qualified Data.ByteString.Base64 as BS
 import qualified Data.ByteString.Lazy as LBS
 import Data.Default.Class
-import Data.Foldable (any)
 import Data.Time
 
 import Network.Wai.SAML2.XML.Encrypted