+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* Added `decodeResponse` and `validateSAMLResponse`
+* Split `validateResponse` into `decodeResponse` and `validateSAMLResponse` ([#31](https://github.com/mbg/wai-saml2/pull/31) by [@fumieval](https://github.com/fumieval))
 * Exported `NameID` (formerly `NameId`), and renamed `subjectNameId` to `subjectNameID`
 
 ## 0.3