+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Split `validateResponse` into `decodeResponse` and `validateSAMLResponse` ([#31](https://github.com/mbg/wai-saml2/pull/31) by [@fumieval](https://github.com/fumieval))
 * Exported `NameID` (formerly `NameId`), and renamed `subjectNameId` to `subjectNameID`
 * Support GHC 9.4 ([#36](https://github.com/mbg/wai-saml2/pull/36) by [@mbg](https://github.com/mbg))